moved some modules around into different sub-packages
