+++ arcana/exceptions.py
