+++ arcana/repository/directory.py
@@ -9,8 +9,8 @@ import logging
 import json
 from fasteners import InterProcessLock
 from arcana.data import Fileset, Field
-from arcana.provenance import Record
-from arcana.exception import (
+from arcana.pipeline.provenance import Record
+from arcana.exceptions import (
     ArcanaError, ArcanaUsageError,
     ArcanaBadlyFormattedDirectoryRepositoryError,
     ArcanaMissingDataException)
