+++ arcana/utils/resources/bash/antsRegistrationSyN.sh
