+++ arcana/utils/resources/bash/check_field_map.sh
