+++ arcana/utils/resources/bash/optiBET.sh
