+++ arcana/utils/resources/bash/zip.sh
